ex01 works well most(all?) edge cases covered - not clear yet if res should be double or float?

+++ CPP_09/ex01/incl/RPN.hpp
@@ -28,7 +28,7 @@ class RPN {
 private:
 	std::list<char>	_rpn_data;
 	std::list<char>	_norm_data;
-	int 			_res;
+	long 			_res;
 	std::string 	_input;
 
 	int 			parse_input();
